requirements.txt
streamlit
whoosh
faiss-cpu
sentence-transformers
numpy
openai
python-dotenv